@@ -1,18 +1,11 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
+namespace ZXBox.Hardware.Input;
 
-namespace ZXBox.Hardware.Input
+public enum JoystickTypeEnum
 {
-    public enum JoystickTypeEnum
-    {
-        None,
-        Sinclair1,
-        Sinclair2,
-        Cursor,
-        Kempston
-        
-        
-    }
+    None,
+    Sinclair1,
+    Sinclair2,
+    Cursor,
+    Kempston
+
 }
