@@ -1,9 +1,6 @@
-using System;
+namespace ZXBox.Hardware.Interfaces;
 
-namespace ZXBox.Hardware.Interfaces
+public interface IInput
 {
-    public interface IInput
-    {
-        int Input(int Port,int tact);
-    }
+    int Input(int Port, int tact);
 }
